Move trait provider spec in the right place
