+++ modules/core/test/resources/META-INF/smithy/manifest
@@ -1,0 +1,1 @@
+traits.smithy
