+++ modules/protocol-tests/resources/META-INF/smithy/manifest
@@ -8,4 +8,3 @@ Health.smithy
 Pizza.smithy
 RoundTrip.smithy
 Version.smithy
-traits.smithy
